made the dependency on Boost::Serialization optional and disabled by default added --enable-serialization (default no) and --enable-python (default yes) configure options fixed compiler flags when --enable-debug is selected added #include "config.h" to all .cc files minor fixes to compile the code with gcc 4.2 in UserFunction and Attribute added boost.m4 macro package to improve checking for boost libraries in configure

+++ src/interval.cc
@@ -104,7 +104,7 @@ Interval trunc(Interval x)
 
 Interval trunc(Interval x, uint n)
 {
-    Interval e=Interval(std::pow(2.0,52-n));
+    Interval e=Interval(std::pow(2.0,52-(int)n));
     Interval y=x+e;
     return y-e;
 }
